refactor: move ScreenEvent to a new file
